Fixed complex norms

upgrade to master

remove warnings

gram schmidt warning removed

gitignore fix

gitignorefix

gitignorefix

unglkfull fix

unglkfull fix

unglkfull fix

unglkfull fix

m-patches + helmholtz_contrast

m-patches stuff

m-patches stuff

randrangefinder + powerits

poweritevamp

poweritmpfix

rrf gram_schmidt added

added extend to rrf

added extend to rrf2

testm+

dels

rand

[algorithms.randrangefinder] added randomized algorithms

[vectorarrays.interfaces] complex fix in norm/norm2

lmor

del MA

ipynb

[lmor]

+++ src/pymor/algorithms/gram_schmidt.py
@@ -59,7 +59,7 @@ def gram_schmidt(A, product=None, return_R=False, atol=1e-13, rtol=1e-13, offset
         A = A.copy()
 
     # main loop
-    R = np.eye(len(A))
+    R = np.eye(len(A), dtype = A.data.dtype)
     remove = []  # indices of to be removed vectors
     for i in range(offset, len(A)):
         # first calculate norm
